chore: dependency updates, see note
This is a temporary fix for the dependency issues we've been having. Technically the source of all these issues, is a utility to calculate the multihash of files. It wasn't used by anything, so really there's no point in keeping it around for now at least.

Ultimately, I would like to reintroduce this utility overtime, however in the short-term it makes sense to remove it.